Refactor Autoload and Namespace Structure; Add Collection Unit Tests and Data Provider
**Details:**
- **Namespace Updates:**
  - Renamed `FireHub\Tests\Unit\Domain\Autoload` namespace to `FireHub\Tests\Unit\Support` for consistency with the test directory structure.
  - Adjusted relevant `@package` and file metadata annotations for clarity.

- **Classmap Refinement:**
  - Removed unused classmap entries for `Collection2` and `Collection3` from `classmap.php` and `firehub.CompiledClassmap.php` to clean up and streamline the autoload definitions.

- **Shared Trait Enhancements:**
  - Updated `each` method in `firehub.Shared` with additional exceptions:
    - `WrongReturnTypeException` to enforce stricter validation of callback return signals.
    - Generic `Throwable` contract for broader error handling.

- **New Unit Tests:**
  - Added `CollectionTest` to test high-level features of the `Collection` data structure:
    - Includes tests for `toArray`, JSON serialization, and object serialization/deserialization.
    - Leverages modern PHPUnit features, such as `TestWith` attributes for parameterized testing.
  - Incorporated a new data provider class (`DataStructureDataProvider`) for reusable data scaffolding in `CollectionTest`.

- **Miscellaneous:**
  - Simplified `$Id` version annotations across modified files for cleaner metadata.

These changes improve test coverage, streamline autoload behavior, and enhance the modularity and maintainability of FireHub Core's unit testing and namespace organization.
